requirements.txt
dc-qiskit-algorithms==0.0.14
docutils==0.20.1
ipykernel==6.29.2
jinja2==3.1.3
nbsphinx==0.9.3
pybind11==2.11.1
pygments==2.17.2
pygments-github-lexers==0.0.5
qiskit==1.0.2
qiskit-aer==0.14.1
qiskit-ibm-runtime==0.23.0
qiskit-ibm-provider==0.11.0
sphinxcontrib-bibtex==2.6.2
sphinx-automodapi==0.17.0
pennylane-sphinx-theme